+++ cli/loom/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.19.1](https://github.com/Yassimba/loom/compare/loom-v0.19.0...loom-v0.19.1) (2026-08-30)
+
+
+### Bug Fixes
+
+* **release:** publish all released Pi packages ([7f81e8a](https://github.com/Yassimba/loom/commit/7f81e8a37f5766eb769533c219be0d17c86f340e))
+
+## [0.19.0](https://github.com/Yassimba/loom/compare/loom-v0.18.1...loom-v0.19.0) (2026-08-30)
+
+
+### Features
+
+* publish add-dir with persistent scopes ([9bb3d8f](https://github.com/Yassimba/loom/commit/9bb3d8ff508055cd8811405df5b5d003dcac534d))
+
 ## [0.18.1](https://github.com/Yassimba/loom/compare/loom-v0.18.0...loom-v0.18.1) (2026-08-29)